Implement Group and Model

+++ src/frameworks/react/components/objects/Gltf.tsx
@@ -3,6 +3,7 @@ import ObjectManager from 'engine/managers/entities/objects/object/object.manage
 import {forwardRef, useEffect} from 'react';
 import useObjects from '../../hooks/useObjects';
 import {setReactRef} from '../../utils/misc';
+import {useParent} from './Group';
 
 type GltfProps = {
   url: string;
@@ -11,20 +12,26 @@ type GltfProps = {
 
 const Gltf = forwardRef<ObjectManager, GltfProps>((props, ref) => {
   const objects = useObjects();
+  const parent = useParent();
 
   useEffect(() => {
-    const object = objects.createGltf(props.url, props.props);
+    if (parent?.destroyed === true) return;
+
+    const object = objects.createGltf(props.url, {
+      parentId: parent?.id,
+      ...props.props,
+    });
 
     setReactRef(ref, object);
 
     return () => {
       objects.destroy(object);
     };
-  }, [objects, props]);
+  }, [objects, props, parent]);
 
   return null;
 });
 
-Gltf.displayName = 'Line';
+Gltf.displayName = 'Gltf';
 
 export default Gltf;
