+++ src/frameworks/react/components/objects/Line.tsx
@@ -4,6 +4,7 @@ import ObjectManager from 'engine/managers/entities/objects/object/object.manage
 import {forwardRef, useEffect} from 'react';
 import useObjects from '../../hooks/useObjects';
 import {setReactRef} from '../../utils/misc';
+import {useParent} from './Group';
 
 type LineProps = {
   points: Vector3Array[];
@@ -13,16 +14,22 @@ type LineProps = {
 
 const Line = forwardRef<ObjectManager, LineProps>((props, ref) => {
   const objects = useObjects();
+  const parent = useParent();
 
   useEffect(() => {
-    const object = objects.createLine(props.points, props.color, props.props);
+    if (parent?.destroyed === true) return;
+
+    const object = objects.createLine(props.points, props.color, {
+      parentId: parent?.id,
+      ...props.props,
+    });
 
     setReactRef(ref, object);
 
     return () => {
       objects.destroy(object);
     };
-  }, [objects, props]);
+  }, [objects, props, parent]);
 
   return null;
 });
